barcode_primer_design/bartender/requirements.txt
flask
flask-wtf
werkzeug
biopython
reportlab